Rename methods to utils
